refactor(*): shared Vitest exposed config and utilities
ref: MANAGER-18041
