--- packages/manager/core/unit-tests-config/src/unit-tests-api.d.ts
@@ -1,10 +1,0 @@
-export {
-  vi,
-  describe,
-  it,
-  expect,
-  beforeEach,
-  afterEach,
-  afterAll,
-  beforeAll,
-} from 'vitest';
